+++ src/main/kotlin/me/partlysanestudios/partlysaneskies/utils/MinecraftUtils.kt
@@ -13,6 +13,8 @@ import me.partlysanestudios.partlysaneskies.PartlySaneSkies
 import me.partlysanestudios.partlysaneskies.utils.StringUtils.removeColorCodes
 import net.minecraft.client.gui.GuiScreen
 import net.minecraft.client.network.NetworkPlayerInfo
+import net.minecraft.entity.Entity
+import net.minecraft.entity.item.EntityArmorStand
 import net.minecraft.inventory.IInventory
 import net.minecraft.item.ItemStack
 import net.minecraft.scoreboard.ScorePlayerTeam
@@ -194,4 +196,47 @@ object MinecraftUtils {
         }
         return false
     }
+
+    /**
+     * @return all the entities loaded in the world
+     */
+    fun getAllEntitiesInWorld(): List<Entity> {
+        return PartlySaneSkies.minecraft.theWorld.getLoadedEntityList()
+    }
+
+    /**
+     * @return all the pets currently loaded in the world
+     */
+    fun getAllPets(): List<Entity> {
+        val petEntities: MutableList<Entity> = java.util.ArrayList()
+        val armorStandEntities = getAllArmorStands()
+        val pattern = """§8\[§7Lv(\d+)§8] (§\w)([\w']+)\s*('s)? (\w+)""".toRegex()
+
+        // For every armor stand in the game, check if it's pet by looking for the level tag in front of the name.
+        // Ex: "*[Lv*100] Su386's Black Cat"
+        for (entity in armorStandEntities) {
+            if (pattern.find(entity.name) != null) {
+                petEntities.add(entity) // If so, add it to the list
+            }
+        }
+        return petEntities
+    }
+
+    /**
+     * @return all the armor stands currently loded in the world
+     */
+    fun getAllArmorStands(): List<Entity> {
+        val armorStandEntities: MutableList<Entity> = java.util.ArrayList()
+        val allEntities = getAllEntitiesInWorld()
+
+        // For every entity in the world, check if its instance of an armor stand
+        for (entity in allEntities) {
+            if (entity is EntityArmorStand) {
+                armorStandEntities.add(entity) // If so, add it to the list
+            }
+        }
+        return armorStandEntities
+    }
+
 }
+
